+++ src/Schema/migrations/getSyncChanges/index.js
@@ -1,23 +1,38 @@
 // @flow
 
 import { uniq, groupBy, toPairs, piped } from 'rambdax'
-import type { MigrationStep } from '../index'
-import type { TableName, ColumnName } from '../../index'
+import type { CreateTableMigrationStep, AddColumnsMigrationStep, SchemaMigrations } from '../index'
+import type { TableName, ColumnName, SchemaVersion } from '../../index'
+import { tableName } from '../../index'
+import { stepsForMigration } from '../stepsForMigration'
 
 import { invariant } from '../../../utils/common'
 import { unnest } from '../../../utils/fp'
 
 export type MigrationSyncChanges = $Exact<{
+  +from: SchemaVersion,
   +tables: TableName<any>[],
   +columns: $Exact<{
     table: TableName<any>,
     columns: ColumnName[],
   }>[],
-}>
+}> | null
 
-// TODO: if we have more than these two step types, it's safer if we take SchemaMigrations and from/to
-// to ensure we process steps in order
-export default function getSyncChanges(steps: MigrationStep[]): MigrationSyncChanges {
+export default function getSyncChanges(
+  migrations: SchemaMigrations,
+  fromVersion: SchemaVersion,
+  toVersion: SchemaVersion,
+): MigrationSyncChanges {
+  const steps = stepsForMigration({ migrations, fromVersion, toVersion })
+  invariant(steps, 'Necessary range of migrations for sync is not available')
+  invariant(
+    toVersion === migrations.maxVersion,
+    'getSyncChanges toVersion should be equal to maxVersion of migrations',
+  )
+  if (fromVersion === toVersion) {
+    return null
+  }
+
   steps.forEach(step => {
     invariant(
       ['create_table', 'add_columns'].includes(step.type),
@@ -25,21 +40,28 @@ export default function getSyncChanges(steps: MigrationStep[]): MigrationSyncCha
     )
   })
 
-  const createdTables = steps
-    .filter(step => step.type === 'create_table')
-    .map(step => step.schema.name)
+  // $FlowFixMe
+  const createTableSteps: CreateTableMigrationStep[] = steps.filter(
+    step => step.type === 'create_table',
+  )
+  const createdTables = createTableSteps.map(step => step.schema.name)
 
-  const allAddedColumns = steps
-    .filter(step => step.type === 'add_columns' && !createdTables.includes(step.table))
+  // $FlowFixMe
+  const addColumnSteps: AddColumnsMigrationStep[] = steps.filter(
+    step => step.type === 'add_columns',
+  )
+  const allAddedColumns = addColumnSteps
+    .filter(step => !createdTables.includes(step.table))
     .map(({ table, columns }) => columns.map(({ name }) => ({ table, name })))
 
   const columnsByTable = piped(allAddedColumns, unnest, groupBy(({ table }) => table), toPairs)
   const addedColumns = columnsByTable.map(([table, columnDefs]) => ({
-    table,
+    table: tableName(table),
     columns: uniq(columnDefs.map(({ name }) => name)),
   }))
 
   return {
+    from: fromVersion,
     tables: uniq(createdTables),
     columns: addedColumns,
   }