+++ src/Schema/migrations/getSyncChanges/test.js
@@ -1,24 +1,28 @@
 import getSyncChanges from './index'
-import { createTable, addColumns } from '../index'
+import { schemaMigrations, createTable, addColumns } from '../index'
 
 const createCommentsTable = createTable({
   name: 'comments',
   columns: [{ name: 'post_id', type: 'string', isIndexed: true }, { name: 'body', type: 'string' }],
 })
 
+const test = (migrations, from, to) => getSyncChanges(schemaMigrations({ migrations }), from, to)
+const testSteps = steps =>
+  getSyncChanges(schemaMigrations({ migrations: [{ toVersion: 2, steps }] }), 1, 2)
+
 describe('getSyncChanges', () => {
+  it('returns null for from==to', () => {
+    expect(test([{ toVersion: 2, steps: [createCommentsTable] }], 2, 2)).toEqual(null)
+  })
   it('returns empty changes for empty steps', () => {
-    expect(getSyncChanges([])).toEqual({ tables: [], columns: [] })
+    expect(testSteps([])).toEqual({ from: 1, tables: [], columns: [] })
   })
   it('returns created tables', () => {
-    expect(getSyncChanges([createCommentsTable])).toEqual({
-      tables: ['comments'],
-      columns: [],
-    })
+    expect(testSteps([createCommentsTable])).toEqual({ from: 1, tables: ['comments'], columns: [] })
   })
   it('returns added columns', () => {
     expect(
-      getSyncChanges([
+      testSteps([
         addColumns({
           table: 'posts',
           columns: [
@@ -28,13 +32,14 @@ describe('getSyncChanges', () => {
         }),
       ]),
     ).toEqual({
+      from: 1,
       tables: [],
       columns: [{ table: 'posts', columns: ['subtitle', 'is_pinned'] }],
     })
   })
   it('combines added columns from multiple migration steps', () => {
     expect(
-      getSyncChanges([
+      testSteps([
         addColumns({
           table: 'posts',
           columns: [{ name: 'subtitle', type: 'string', isOptional: true }],
@@ -49,26 +54,31 @@ describe('getSyncChanges', () => {
         }),
       ]),
     ).toEqual({
+      from: 1,
       tables: [],
       columns: [{ table: 'posts', columns: ['subtitle', 'is_pinned', 'author_id'] }],
     })
   })
   it('skips added columns for a table if it is also added', () => {
     expect(
-      getSyncChanges([
+      testSteps([
         createCommentsTable,
         addColumns({
           table: 'comments',
           columns: [{ name: 'reactions', type: 'string', isOptional: true }],
         }),
       ]),
-    ).toEqual({ tables: ['comments'], columns: [] })
+    ).toEqual({
+      from: 1,
+      tables: ['comments'],
+      columns: [],
+    })
   })
   it('skips duplicates', () => {
     // technically, a duplicate createTable or addColumn would crash
     // but this is in case future migration types could do something like it
     expect(
-      getSyncChanges([
+      testSteps([
         createCommentsTable,
         createCommentsTable,
         addColumns({
@@ -81,17 +91,30 @@ describe('getSyncChanges', () => {
         }),
       ]),
     ).toEqual({
+      from: 1,
       tables: ['comments'],
       columns: [{ table: 'posts', columns: ['subtitle'] }],
     })
   })
-  it('can handle a complex migration steps list', () => {
-    expect(
-      getSyncChanges([
+  const bigMigrations = [
+    {
+      toVersion: 10,
+      steps: [
+        // No changes
+      ],
+    },
+    {
+      toVersion: 9,
+      steps: [
         addColumns({
           table: 'attachment_versions',
           columns: [{ name: 'reactions', type: 'string' }],
         }),
+      ],
+    },
+    {
+      toVersion: 8,
+      steps: [
         addColumns({
           table: 'workspaces',
           columns: [
@@ -99,10 +122,20 @@ describe('getSyncChanges', () => {
             { name: 'limits', type: 'string' },
           ],
         }),
+      ],
+    },
+    {
+      toVersion: 7,
+      steps: [
         createTable({
           name: 'attachments',
           columns: [{ name: 'parent_id', type: 'string', isIndexed: true }],
         }),
+      ],
+    },
+    {
+      toVersion: 6,
+      steps: [
         createTable({
           name: 'attachment_versions',
           columns: [
@@ -115,37 +148,74 @@ describe('getSyncChanges', () => {
             { name: 'created_at', type: 'number' },
           ],
         }),
+      ],
+    },
+    {
+      toVersion: 5,
+      steps: [
         addColumns({
           table: 'comments',
           columns: [{ name: 'is_pinned', type: 'boolean' }, { name: 'extra', type: 'string' }],
         }),
-        addColumns({
-          table: 'projects',
-          columns: [{ name: 'extra', type: 'string' }],
-        }),
+        addColumns({ table: 'projects', columns: [{ name: 'extra', type: 'string' }] }),
+      ],
+    },
+    { toVersion: 4, steps: [] },
+    {
+      toVersion: 3,
+      steps: [
         addColumns({
           table: 'task_recurrences',
           columns: [{ name: 'project_id', type: 'string' }],
         }),
+      ],
+    },
+    {
+      toVersion: 2,
+      steps: [
         addColumns({
           table: 'projects',
           columns: [{ name: 'preferences', type: 'string', isOptional: true }],
         }),
-      ]),
-    ).toEqual({
-      tables: ['attachments', 'attachment_versions'],
+      ],
+    },
+  ]
+  it('can handle a complex migration steps list', () => {
+    expect(test(bigMigrations, 1, 10)).toEqual({
+      from: 1,
+      tables: ['attachment_versions', 'attachments'],
       columns: [
-        { table: 'workspaces', columns: ['plan_info', 'limits'] },
-        { table: 'comments', columns: ['is_pinned', 'extra'] },
-        { table: 'projects', columns: ['extra', 'preferences'] },
+        { table: 'projects', columns: ['preferences', 'extra'] },
         { table: 'task_recurrences', columns: ['project_id'] },
+        { table: 'comments', columns: ['is_pinned', 'extra'] },
+        { table: 'workspaces', columns: ['plan_info', 'limits'] },
       ],
     })
+  })
+  it(`returns only the necessary range of migrations`, () => {
+    expect(test(bigMigrations, 6, 10)).toEqual({
+      from: 6,
+      tables: ['attachments'],
+      columns: [
+        { table: 'workspaces', columns: ['plan_info', 'limits'] },
+        { table: 'attachment_versions', columns: ['reactions'] },
+      ],
+    })
+    expect(test(bigMigrations, 8, 10)).toEqual({
+      from: 8,
+      tables: [],
+      columns: [{ table: 'attachment_versions', columns: ['reactions'] }],
+    })
+    expect(test(bigMigrations, 9, 10)).toEqual({ from: 9, tables: [], columns: [] })
+    expect(test(bigMigrations, 10, 10)).toEqual(null)
+  })
+  it(`fails on incorrect migrations`, () => {
+    expect(() => test(bigMigrations, 0, 9)).toThrow()
+    expect(() => test(bigMigrations, 8, 11)).toThrow()
   })
   it('fails early on unknown migration steps', () => {
     const possibleFutureTypes = [
       'broken',
-      undefined,
       'rename_table',
       'rename_column',
       'add_column_index',
@@ -155,7 +225,8 @@ describe('getSyncChanges', () => {
       'destroy_column',
     ]
     possibleFutureTypes.forEach(type => {
-      expect(() => getSyncChanges([{ type }])).toThrow(/Unknown migration step type/)
+      expect(() => testSteps([{ type }])).toThrow(/Unknown migration step type/)
     })
+    expect(() => testSteps([{ type: undefined }])).toThrow(/Invalid migration steps/)
   })
 })