+++ CHANGELOG.md
@@ -6,6 +6,11 @@ All notable changes to this project will be documented in this file.
 
 ### New features
 
+- [Sync] Introducing Migration Syncs - this allows fully consistent synchronization when migrating
+      between schema versions. Previously, there was no mechanism to incrementally fetch all remote changes in
+      new tables and columns after a migration - so local copy was likely inconsistent, requiring a re-login.
+      After adopting migration syncs, Watermelon Sync will request from backend all missing information.
+      See Sync docs for more details.
 - [iOS] Introducing a new native SQLite database integration, rewritten from scratch in C++, based
        on React Native's JSI (JavaScript Interface). It is to be considered experimental, however
        we intend to make it the default (and eventually, the only) implementation. In a later release,
@@ -15,6 +20,7 @@ All notable changes to this project will be documented in this file.
        however this speedup is only achieved with some unpublished React Native patches.
 
        To try out JSI, add `experimentalUseJSI: true` to `SQLiteAdapter` constructor.
+- [Query] Added `Q.experimentalSortBy(sortColumn, sortOrder)`, `Q.experimentalTake(count)`, `Q.experimentalSkip(count)` methods - @Kenneth-KT
 
 ### Changes
 
@@ -28,12 +34,14 @@ All notable changes to this project will be documented in this file.
 - [Sync] Sync no longer fails if pullChanges returns collections that don't exist on the frontend - shows a warning instead. This is to make building backwards-compatible backends less error-prone
 - [Sync] [Docs] Sync documentation has been rewritten, and is now closer in detail to a formal specification
 - [Hardening] database.collections.get() better validates passed value
+- [Hardening] Prevents unsafe strings from being passed as column name/table name arguments in QueryDescription
 
 ### Fixes
 
 - [iOS] Fixed a bug that could cause a database operation to fail with an (6) SQLITE_LOCKED error
 - [iOS] Fixed 'jsi/jsi.h' file not found when building at the consumer level. Added path `$(SRCROOT)/../../../../../ios/Pods/Headers/Public/React-jsi` to Header Search Paths (issue #691)
 - [Native] SQLite keywords used as table or column names no longer crash
+- Fixed potential issues when subscribing to database, collection, model, queries passing a subscriber function with the same identity more than once
 
 ### Internal